Refactor CMakeLists to be much cleaner
* Make building as a shared library on MSVC work for haicrypt
* Up the CMake requirement to 3.0.2
* Add all arbitrary options as offical cmake options
