+++ include/win/wintime.h
@@ -4,6 +4,7 @@
 #include <winsock2.h>
 #include <windows.h>
 #include <time.h>
+#include "haicrypt.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -33,7 +34,7 @@ struct timezone
 
 void timeradd(struct timeval *a, struct timeval *b, struct timeval *result);
 
-int gettimeofday(struct timeval* tp, struct timezone* tz);
+HAICRYPT_API int gettimeofday(struct timeval* tp, struct timezone* tz);
 
 #ifdef __cplusplus
 }
